Separated contents to the corresponding file
